Rename /api/ subpackages to /controller/
