+++ api/src/main/java/com/example/echo_api/modules/clerk/controller/ClerkDevController.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.modules.clerk.api;
+package com.example.echo_api.modules.clerk.controller;
 
 import java.util.List;
