+++ api/src/main/java/com/example/echo_api/modules/profile/controller/ProfileManagementController.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.modules.profile.api;
+package com.example.echo_api.modules.profile.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
