+++ api/src/test/java/com/example/echo_api/modules/post/controller/PostManagementControllerIT.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.modules.post.api;
+package com.example.echo_api.modules.post.controller;
 
 import java.util.UUID;
 import java.util.stream.Stream;
