+++ api/src/test/java/com/example/echo_api/modules/profile/controller/ProfileViewControllerTest.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.modules.profile.api;
+package com.example.echo_api.modules.profile.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
